pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bilix"
dynamic = ["version"]
description = "⚡️Lightning-fast asynchronous download tool for bilibili and more"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
    { name = "HFrost0", email = "hhlfrost@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "aiofiles>=0.8.0",
    "anyio",
    "danmakuC>=0.3.4",
    "bs4",
    "click>=8.0.3",
    "httpx[http2]>=0.23.3",
    "json5",
    "m3u8",
    "pycryptodome",
    "pydantic",
    "pyexecjs",
    "rich",
    "browser_cookie3>=0.17.1",
    "pymp4>=1.2.0",
]

[project.optional-dependencies]
serve = [
    "fastapi>=0.92.0",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "starlette",
]

[project.scripts]
bilix = "bilix.cli.main:main"

[project.urls]
Homepage = "https://github.com/HFrost0/bilix"

[tool.hatch.version]
path = "bilix/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/bilix",
]
